Separate logical parts of component into hooks
